Add working kvm sectorlisp
Although it still crashes with MMIO error on bad input (which may be
suppressable behavior), the issue from last commit -- it crashing
immeadiately with an MMIO error -- was solved by fixing a typo in the
sed patch. Now, a forked repo will be used for sectorlisp with the kvm
changes to avoid sed.